+++ config/Migrations/20201014000227_MigrationTimezones.php
@@ -57,7 +57,7 @@ class MigrationTimezones extends AbstractMigration {
 				'precision' => 10,
 				'scale' => 6,
 			])
-			->addColumn('lng', 'float', [
+			->addColumn('lng', 'decimal', [
 				'default' => null,
 				'null' => true,
 				'precision' => 10,
